changed createTransfer path to /transfer. Combined getAccountByUserId and getAccountByAccountId methods in AccountController to not confuse Spring. Corrected >= to <= in compareTo method in App.sendBucks(). Transfer is working! not able to create one in postman yet.

+++ tenmo-client/src/main/java/com/techelevator/tenmo/model/Transfer.java
@@ -51,19 +51,19 @@ public class Transfer {
         this.transferStatusId = transferStatusId;
     }
 
-    public Long getAccountFrom() {
+    public int getAccountFrom() {
         return accountFrom;
     }
 
-    public void setAccountFrom(Long accountFrom) {
+    public void setAccountFrom(int accountFrom) {
         this.accountFrom = accountFrom;
     }
 
-    public Long getAccountTo() {
+    public int getAccountTo() {
         return accountTo;
     }
 
-    public void setAccountTo(Long accountTo) {
+    public void setAccountTo(int accountTo) {
         this.accountTo = accountTo;
     }
